# src/backend/TestPlanService/Api/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "127.0.0.1:5000",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/projects/{projectId}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "Projects_DeleteProject",
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TestSuites"
        ],
        "operationId": "TestSuites_MoveSuite",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "id",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "moveSuiteRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MoveSuiteRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/runs": {
      "post": {
        "tags": [
          "TestRuns"
        ],
        "operationId": "TestRuns_Add",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddRunRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "/projects/{projectId}/runs/{runId}/results": {
      "post": {
        "tags": [
          "TestRuns"
        ],
        "operationId": "TestRuns_AddResult",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "runId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TestResultRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "get": {
        "tags": [
          "TestRuns"
        ],
        "operationId": "TestRuns_GetResults",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "runId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TestRunPoint"
              }
            }
          }
        }
      }
    },
    "/projects/{projectId}/runs/{runId}/results/{resultId}/files": {
      "post": {
        "tags": [
          "TestRuns"
        ],
        "operationId": "TestRuns_AddFiles",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "projectId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "resultId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "ids",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            },
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "runId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserIdentity": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "Outcome": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Unknow",
        "Planed",
        "Passed",
        "Failed",
        "Blocked",
        "Skipped",
        "Paused"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ]
    },
...
    "AddRunRequest": {
      "type": "object",
      "required": [
        "isAutomated"
      ],
      "properties": {
        "isAutomated": {
          "type": "boolean"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "TestResultRequest": {
      "type": "object",
      "required": [
        "startedDate",
        "outCome",
        "completedDate"
      ],
      "properties": {
        "automatedTestId": {
          "type": "string"
        },
        "automatedTestName": {
          "type": "string"
        },
        "automatedTestStorage": {
          "type": "string"
        },
        "computerName": {
          "type": "string"
        },
        "startedDate": {
          "type": "string",
          "format": "date-time"
        },
        "errorMessage": {
          "type": "string"
        },
        "outCome": {
          "$ref": "#/definitions/Outcome"
        },
        "stackTrace": {
          "type": "string"
        },
        "completedDate": {
          "type": "string",
          "format": "date-time"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "TestRunPoint": {
      "type": "object",
      "required": [
        "id",
        "startedDate",
        "outCome",
        "completedDate"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "automatedTestId": {
          "type": "string"
        },
        "automatedTestName": {
          "type": "string"
        },
        "automatedTestStorage": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "computerName": {
          "type": "string"
        },
        "startedDate": {
          "type": "string",
          "format": "date-time"
        },
        "errorMessage": {
          "type": "string"
        },
        "outCome": {
          "$ref": "#/definitions/Outcome"
        },
        "owner": {
          "$ref": "#/definitions/UserIdentity"
        },
        "stackTrace": {
          "type": "string"
        },
        "completedDate": {
          "type": "string",
          "format": "date-time"
        },
        "files": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TestPointFile"
          }
        }
      }
    },
    "TestPointFile": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "MoveSuiteRequest": {
      "type": "object",
      "required": [
        "newParentId"
      ],
      "properties": {
        "newParentId": {
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }
}